date-picker: Split `DateRangePicker` into own entrypoint (#1163)
