+++ .storybook/stories/FormFields.tsx
@@ -5,7 +5,8 @@ import { Button, ButtonGroup } from '@ag.ds-next/react/button';
 import { Columns, Column } from '@ag.ds-next/react/columns';
 import { Combobox, ComboboxMulti } from '@ag.ds-next/react/src/combobox';
 import { ControlGroup, Checkbox, Radio } from '@ag.ds-next/react/control-input';
-import { DatePicker, DateRangePicker } from '@ag.ds-next/react/date-picker';
+import { DatePicker } from '@ag.ds-next/react/date-picker';
+import { DateRangePicker } from '@ag.ds-next/react/date-range-picker';
 import { FileInput } from '@ag.ds-next/react/file-input';
 import { FileUpload } from '@ag.ds-next/react/file-upload';
 import { FormStack } from '@ag.ds-next/react/form-stack';
