+++ example-site/components/templates/SinglePageForm.tsx
@@ -22,7 +22,8 @@ import { useScrollToField } from '@ag.ds-next/react/field';
 import { PageContent } from '@ag.ds-next/react/content';
 import { Column, Columns } from '@ag.ds-next/react/columns';
 import { Breadcrumbs } from '@ag.ds-next/react/breadcrumbs';
-import { DatePicker, DateRangePicker } from '@ag.ds-next/react/date-picker';
+import { DatePicker } from '@ag.ds-next/react/date-picker';
+import { DateRangePicker } from '@ag.ds-next/react/date-range-picker';
 import { Text } from '@ag.ds-next/react/text';
 import { FormDivider } from '../FormDivider';
 import { PageTitle } from '../PageTitle';
