+++ packages/react/date-range-picker/package.json
@@ -1,0 +1,3 @@
+{
+	"main": "dist/ag.ds-next-react-date-range-picker.cjs.js"
+}
