+++ packages/react/src/date-picker/docs/code.mdx
@@ -1,5 +1,3 @@
 ## Props
 
 <ComponentPropsTable name="DatePicker" />
-
-<ComponentPropsTable name="DateRangePicker" />
